+++ src/verify-request/types.ts
@@ -7,6 +7,7 @@ export interface Routes {
 
 type VerifyRequestOptions = {
   accessMode: AccessMode;
+  returnHeader: boolean;
 }
 
 export type Options = Partial<VerifyRequestOptions> & Partial<Routes>;